wip add capability to generate presigned urls using passed-in date
Bookshelf needs the ability to do sigv4 presigned url request verification,
i.e. while processing a sigv4 presigned url request, determine whether the
presigned url is valid, appropriately signed, etc.

The only way I know to do this currently is to create a new presigned url
using whatever inputs the original presigned url was created with, and
compare the two.  In order to do this, I have to use the same date the
original presigned url was created with.  But it wasn't possible to pass
in such a date, thus this commit.

Signed-off-by: Lincoln Baker <lbaker@chef.io>