more consistent treatment of host header
1) host headers are treated inconsistently betw various componenents
(erlcloud, mini_s3, bookshelf, etc).

2) this generated host header was incorrect in certain cases, eg a host
of http://127.0.0.1:4321 left off the port and generated a host header
of 127.0.0.1. however host header specs say that in case of a missing
port the assumed port will be based on the scheme - 80 for http, 443
for https.  so the port should be added in this case (as it is 4321).

https://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html

Signed-off-by: Lincoln Baker <lbaker@chef.io>